# web/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_courses__course_id__documents_post": {
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "purpose": {
            "$ref": "#/components/schemas/DocumentPurpose",
            "default": "MATERIAL"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_courses__course_id__documents_post",
        "type": "object"
      },
...
      "ChunkRead": {
        "description": "A source passage for \"View source\" (docs/PROJECT_SPEC.md \u00a766).",
        "properties": {
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "paragraph_end": {
            "title": "Paragraph End",
            "type": "integer"
          },
          "paragraph_start": {
            "title": "Paragraph Start",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "position",
          "page_number",
          "page_end",
          "section",
          "paragraph_start",
          "paragraph_end",
          "text"
        ],
        "title": "ChunkRead",
        "type": "object"
      },
...
      "DocumentKind": {
        "description": "File types the Knowledge Repository accepts (docs/PROJECT_SPEC.md \u00a716).",
        "enum": [
          "PDF",
          "TEXT",
          "MARKDOWN",
          "DOCX",
          "PPTX",
          "IMAGE"
        ],
        "title": "DocumentKind",
        "type": "string"
      },
      "DocumentPurpose": {
        "description": "What an upload is. MATERIAL is study material the AI builds a curriculum from; a\nQUESTION_BANK holds the user's own labelled questions and expected answers, imported as they\nare, without AI.",
        "enum": [
          "MATERIAL",
          "QUESTION_BANK"
        ],
        "title": "DocumentPurpose",
        "type": "string"
      },
      "DocumentRead": {
        "properties": {
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "chapter_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          },
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Notice"
          },
          "kind": {
            "$ref": "#/components/schemas/DocumentKind"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "purpose": {
            "$ref": "#/components/schemas/DocumentPurpose"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source_created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Created At"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "course_id",
          "chapter_id",
          "filename",
          "mime_type",
          "kind",
          "purpose",
          "size_bytes",
          "sha256",
          "status",
          "error_message",
          "import_notice",
          "page_count",
          "chunk_count",
          "source_created_at",
          "analyzed_at",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentRead",
        "type": "object"
      },
      "DocumentStatus": {
        "description": "Ingestion pipeline state (docs/PROJECT_SPEC.md \u00a717). FAILED documents stay stored, with a\nuser-safe reason, so nothing the user uploaded silently disappears.",
        "enum": [
          "PROCESSING",
          "READY",
          "FAILED"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentUpdate": {
        "additionalProperties": false,
        "description": "Files the document under a Chapter of the same Course, or unassigns it (null). Moving it\nmarks it not analyzed, so its material is analyzed again in its new Chapter.",
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Id"
          }
        },
        "required": [
          "chapter_id"
        ],
        "title": "DocumentUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Adaptive AI Learning Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/documents": {
      "get": {
        "operationId": "list_documents_api_v1_courses__course_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chapter_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chapter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentRead"
                  },
                  "title": "Response List Documents Api V1 Courses  Course Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      },
      "post": {
        "description": "Stores the file and returns immediately (202) with status PROCESSING; poll\nGET /documents/{id} until it is READY or FAILED. Optional form field `chapter_id` files it\nunder a Chapter of this Course.\n\nForm field `purpose`: MATERIAL (default) is study material for curriculum generation.\nQUESTION_BANK holds the user's own questions and expected answers, labelled \"Domanda:\" /\n\"Risposta:\" (or \"Question:\" / \"Answer:\"). Each pair becomes a NOT_STUDIED Concept with one\nLearning Item, under `chapter_id` or a new Chapter named after the file.",
        "operationId": "upload_document_api_v1_courses__course_id__documents_post",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_courses__course_id__documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "update_document_api_v1_documents__document_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/chunks": {
      "get": {
        "operationId": "list_chunks_api_v1_documents__document_id__chunks_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChunkRead"
                  },
                  "title": "Response List Chunks Api V1 Documents  Document Id  Chunks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Chunks",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/chunks/{chunk_id}": {
      "get": {
        "operationId": "get_chunk_api_v1_documents__document_id__chunks__chunk_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "chunk_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Chunk Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Chunk",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{document_id}/file": {
      "get": {
        "description": "The original file, to its owner only (404 for anyone else, as for a missing id). Always\nan attachment, never rendered inline, so an uploaded HTML or SVG can't run in the app's\norigin. There are no public or pre-signed URLs: every download passes this check.",
        "operationId": "download_document_api_v1_documents__document_id__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Document",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}